WIP: H3m map reader - heroes details

+++ pkg/h3m/Map.go
@@ -9,7 +9,7 @@ type H3m struct {
 	HasUnderground bool
 	Name           string
 	Description    string
-	Difficulty     uint32
+	Difficulty     byte
 
 	Players []*models.Player
 }
